requirements.txt
Django==1.4.5
httplib2==0.8
oauth2==1.5.211
python-twitter==0.8.5
simplejson==3.1.2
wsgiref==0.1.2

